updated highscorelist

+++ app/lib/HighscoreHandler.php
@@ -1,4 +1,4 @@
-<?php 
+<?php
 
 require_once "DbConnect.php";
 
@@ -30,26 +30,57 @@ class HighscoreHandler extends DbConnect {
 			if($diff > HS_DELAY)
 			{
 				$this->_write();
-			}	
+			}
 		}
+	}
+
+	public function listHighscores()
+	{
+		echo "<table>";
+		echo "<tr>";
+		echo "<td>Username</td><td>Level</td><td>Points</td><td>Location</td>";
+		echo "</tr>";
+		$result = $this->db->query("SELECT * FROM `highscores`");
+		while($row = $result->fetch_object())
+		{
+			echo "<tr>";
+			echo "<td>" . $row->username . "</td>";
+			echo "<td>" . $row->level . "</td>";
+			echo "<td>" . $row->points . "</td>";
+			echo "<td>" . $row->location . "</td>";
+			echo "</tr>";
+		}
+		echo "</table>";
+		exit;
 	}
 
 	private function _write()
 	{
-		if($this->points >= $this->lastRow->points)
+		if($this->lastRow !== null)
 		{
-			$this->db->query("DELETE FROM `highscores` 
-				WHERE id = ". $this->lastRow->id .";");
+			if($this->points >= $this->lastRow->points)
+			{
+				$this->db->query("DELETE FROM `highscores`
+					WHERE id = ". $this->lastRow->id .";");
 
-			$this->db->query("DELETE FROM `highscores` 
-				WHERE ip = '" . $this->ip . "';");
+				$this->db->query("DELETE FROM `highscores`
+					WHERE ip = '" . $this->ip . "';");
+			}
+
+			if($this->points > $this->lastRow->points)
+			{
+				$this->db->query("INSERT INTO `highscores`
+					(username, level, points, location, ip)
+						VALUES
+					('" . $this->username . "', '" . $this->level . "', '" . $this->points . "', '" . $this->location . "', '" . $this->ip . "');");
+			}
 		}
-
-		//if($this->lastrow <= $this->nowrow)
-
-		$this->db->query("INSERT INTO `highscores` 
-			(username, level, points, location, ip) 
-				VALUES 
-			('" . $this->username . "', '" . $this->level . "', '" . $this->points . "', '" . $this->location . "', '" . $this->ip . "');");			
+		else
+		{
+			$this->db->query("INSERT INTO `highscores`
+					(username, level, points, location, ip)
+						VALUES
+					('" . $this->username . "', '" . $this->level . "', '" . $this->points . "', '" . $this->location . "', '" . $this->ip . "');");
+		}
 	}
 }
